+++ lib/mountain_menu.dart
@@ -1,9 +1,9 @@
 library mountain_menu;
 
-
 import 'package:flutter/material.dart';
 import 'package:mountain_menu/arc.dart';
 import 'package:mountain_menu/arc_border.dart';
+import 'package:mountain_menu/positions.dart';
 
 class MountainMenu extends StatefulWidget {
   final Color color;
@@ -17,6 +17,7 @@ class MountainMenu extends StatefulWidget {
   final double? borderWidth;
   final Duration? duration;
   final Color? borderColor;
+  final bool mini;
 
   const MountainMenu(
       {Key? key,
@@ -24,6 +25,7 @@ class MountainMenu extends StatefulWidget {
       required this.menuIcon,
       required this.menus,
       required this.onChange,
+      required this.mini,
       this.mainButtonElevation,
       this.borderWidth,
       this.duration,
@@ -43,6 +45,7 @@ class _MountainMenuState extends State<MountainMenu>
     with SingleTickerProviderStateMixin {
   bool _showMenu = false;
   late AnimationController _controller;
+  final Positions _positions = Positions(); //Positions for menus
 
   @override
   void dispose() {
@@ -54,18 +57,18 @@ class _MountainMenuState extends State<MountainMenu>
   void initState() {
     super.initState();
     _controller = AnimationController(
-      duration: widget.duration ?? const Duration(milliseconds: 900),
+      duration: widget.duration ?? const Duration(milliseconds: 700),
       vsync: this,
     );
   }
 
   @override
   Widget build(BuildContext context) {
-    if (widget.menus.length < 3 || widget.menus.length>3) {
-      throw UnimplementedError("Total 3 menus are required!");
+    if (widget.menus.isEmpty || widget.menus.length > 5) {
+      throw UnimplementedError("Menus length is not sufficient!");
     } else {
       return AnimatedContainer(
-        duration: widget.duration ?? const Duration(milliseconds: 700),
+        duration: widget.duration ?? const Duration(milliseconds: 400),
         child: CustomPaint(
           painter: ArcBorder(
               color: widget.borderColor ?? Colors.yellow,
@@ -77,28 +80,24 @@ class _MountainMenuState extends State<MountainMenu>
               width: MediaQuery.of(context).size.width,
               child: Stack(
                 children: [
-
-                  if (_showMenu)
+                  if (_showMenu && widget.menus.length > 1)
                     Positioned(
                       top: 0,
                       left: 0,
                       right: 0,
                       child: SlideTransition(
-                        position: Tween<Offset>(
-                          begin: const Offset(0.0, 0.1),
-                          end: const Offset(0.0, -0.1),
-                        ).animate(CurvedAnimation(
-                          parent: _controller,
-                          curve: widget.curve != null
-                              ? widget.curve as Curve
-                              : Curves.bounceOut,
-                        )),
+                        position:
+                            _positions.topButton().animate(CurvedAnimation(
+                                  parent: _controller,
+                                  curve: widget.curve != null
+                                      ? widget.curve as Curve
+                                      : Curves.easeInOut,
+                                )),
                         child: FloatingActionButton(
                           heroTag: "menuTop",
                           backgroundColor: widget.buttonBackgroundColor,
-                          mini: true,
+                          mini: widget.mini,
                           onPressed: () async {
-
                             widget.onChange(1);
                             _animate();
                           },
@@ -114,7 +113,7 @@ class _MountainMenuState extends State<MountainMenu>
                       heroTag: "mainButton",
                       elevation: widget.mainButtonElevation,
                       backgroundColor: Colors.transparent,
-                      mini: true,
+                      mini: widget.mini,
                       onPressed: () async {
                         _animate();
                       },
@@ -124,22 +123,19 @@ class _MountainMenuState extends State<MountainMenu>
                       ),
                     ),
                   ),
-                  if (_showMenu)
+                  if (_showMenu && widget.menus.isNotEmpty)
                     Positioned(
                       top: MediaQuery.of(context).size.height * .06,
                       left: MediaQuery.of(context).size.width * .2,
                       child: SlideTransition(
-                        position: Tween<Offset>(
-                          begin: const Offset(0.9, 0.0),
-                          end: const Offset(0.4, -0.2),
-                        ).animate(CurvedAnimation(
-                          parent: _controller,
-                          curve: widget.curve != null
-                              ? widget.curve as Curve
-                              : Curves.bounceOut,
-                        )),
+                        position: _positions.halfLeft().animate(CurvedAnimation(
+                              parent: _controller,
+                              curve: widget.curve != null
+                                  ? widget.curve as Curve
+                                  : Curves.easeInOut,
+                            )),
                         child: FloatingActionButton(
-                          mini: true,
+                          mini: widget.mini,
                           heroTag: "menuLeft",
                           backgroundColor: widget.buttonBackgroundColor,
                           onPressed: () async {
@@ -150,30 +146,73 @@ class _MountainMenuState extends State<MountainMenu>
                         ),
                       ),
                     ),
-                  if(_showMenu)
+                  if (_showMenu && widget.menus.length > 2)
                     Positioned(
                       top: MediaQuery.of(context).size.height * .06,
                       right: MediaQuery.of(context).size.width * .2,
                       child: SlideTransition(
-                        position: Tween<Offset>(
-                          begin: const Offset(-0.9, 0.0),
-                          end: const Offset(-0.4, -0.2),
-                        ).animate(CurvedAnimation(
-                          parent: _controller,
-                          curve: widget.curve != null
-                              ? widget.curve as Curve
-                              : Curves.easeOut,
-                        )),
+                        position:
+                            _positions.halfRight().animate(CurvedAnimation(
+                                  parent: _controller,
+                                  curve: widget.curve != null
+                                      ? widget.curve as Curve
+                                      : Curves.easeOut,
+                                )),
                         child: FloatingActionButton(
                           heroTag: "menuRight",
-                          mini: true,
+                          mini: widget.mini,
                           backgroundColor: widget.buttonBackgroundColor,
                           onPressed: () async {
-
                             widget.onChange(2);
                             _animate();
                           },
                           child: widget.menus[2],
+                        ),
+                      ),
+                    ),
+                  if (_showMenu && widget.menus.length > 3)
+                    Positioned(
+                      left: 16,
+                      bottom: 5,
+                      child: SlideTransition(
+                        position: _positions.left().animate(CurvedAnimation(
+                              parent: _controller,
+                              curve: widget.curve != null
+                                  ? widget.curve as Curve
+                                  : Curves.easeOut,
+                            )),
+                        child: FloatingActionButton(
+                          heroTag: "left",
+                          mini: widget.mini,
+                          backgroundColor: widget.buttonBackgroundColor,
+                          onPressed: () async {
+                            widget.onChange(3);
+                            _animate();
+                          },
+                          child: widget.menus[3],
+                        ),
+                      ),
+                    ),
+                  if (_showMenu && widget.menus.length > 4)
+                    Positioned(
+                      right: 16,
+                      bottom: 5,
+                      child: SlideTransition(
+                        position: _positions.right().animate(CurvedAnimation(
+                              parent: _controller,
+                              curve: widget.curve != null
+                                  ? widget.curve as Curve
+                                  : Curves.easeOut,
+                            )),
+                        child: FloatingActionButton(
+                          heroTag: "right",
+                          mini: widget.mini,
+                          backgroundColor: widget.buttonBackgroundColor,
+                          onPressed: () async {
+                            widget.onChange(4);
+                            _animate();
+                          },
+                          child: widget.menus[4],
                         ),
                       ),
                     ),