+++ test/mountain_menu_test.dart
@@ -1,7 +1,4 @@
 import 'package:flutter_test/flutter_test.dart';
-
-import 'package:mountain_menu/mountain_menu.dart';
-
 void main() {
   test('adds one to input values', () {
     // final calculator = Calculator();