+++ CHANGELOG.md
@@ -1,3 +1,3 @@
-## 0.0.1
-
-* TODO: Describe initial release.
+## 0.0.3
+Menus length updated to 5 max.
+You can adjust Menus size to mini or expanded. Mini property is required